// src/utils/cuisineSignatures.generated.ts
export const CUISINE_SIGNATURES = [
  {
    "cuisine": "Hsca",
    "sampleSize": 502,
    "popularityCount": 0,
    "popularityWeight": 0,
    "averageElementals": {
      "Fire": 0.22777347775705398,
      "Water": 0.29187679763580593,
      "Earth": 0.304932965088938,
      "Air": 0.1754167595182021
    },
    "averageAlchemical": {
      "Spirit": 2.5624701195219126,
      "Essence": 3.151494023904383,
      "Matter": 3.3666733067729058,
      "Substance": 3.1024302788844635
    },
    "variance": {
      "elementals": {
        "Fire": 0.06622467525165175,
        "Water": 0.07723037874250549,
        "Earth": 0.0651352221649664,
        "Air": 0.03476266752395651
      },
      "diversityScore": 0.18251470776231013
    },
    "signatures": [],
    "planetaryPatterns": []
  },
// ...
] as const;
export const CUISINE_GLOBAL_AVERAGES = {
  "elementals": {
    "Fire": 0.25630307461274054,
    "Water": 0.2817536332563197,
    "Earth": 0.3378978109461916,
    "Air": 0.12404548118474823
  },
  "alchemical": {
    "Spirit": 3.8206224911671476,
    "Essence": 4.661694759637511,
    "Matter": 5.573986750623534,
    "Substance": 4.877478399904208
  },
  "thermodynamics": {
    "heat": 0,
    "entropy": 0,
    "reactivity": 0,
    "gregsEnergy": 0,
    "kalchm": 0,
    "monica": 0
  },
  "elementalsStdDev": {
    "Fire": 0.06950580523193774,
    "Water": 0.06683760661201944,
    "Earth": 0.07560885601205619,
    "Air": 0.03927935173474447
  },
  "alchemicalStdDev": {
    "Spirit": 3.64332544295432,
    "Essence": 4.622545270317315,
    "Matter": 7.02156374648961,
    "Substance": 5.891884261649369
  },
  "thermodynamicsStdDev": {
    "heat": 0,
    "entropy": 0,
    "reactivity": 0,
    "gregsEnergy": 0,
    "kalchm": 0,
    "monica": 0
  }
} as const;